floats->doubles, across the board

+++ include/debug.h
@@ -2,6 +2,6 @@
 #include "matrix.h"
 #include <stdio.h>
 
-void matrix_print_rowcol(float* arr, int size,int flag);
+void matrix_print_rowcol(double* arr, int size,int flag);
 void matrix_print(const matrix_t *m);
 void matrix_set_linear_range(matrix_t* m);
